// src/components/Table/models/config.models.ts
import { ReactNode } from 'react';

import {
    EColumnDataType,
    IColumnCustomStylingDto,
    TColumnId,
} from '@/components/Table/models/column.models';
import { TDataFieldAccessor } from '@/components/Table/models/data.models';
import { IColumnSortConfig } from '@/components/Table/models/sorting.models';

interface ITableColumnConfig {
    columnDataType?: EColumnDataType; // For default type styling
    label: string;
    columnId: TColumnId;
    dataAccessor: TDataFieldAccessor;
    onCellClick?: () => void;
    renderData?: () => ReactNode;
    customStylingDto?: IColumnCustomStylingDto;
    isSortable?: boolean;
    isAutoGenerated?: boolean;
}

interface ITableConfigProps {
    columnConfigList: ITableColumnConfig[];
    defaultSorting?: IColumnSortConfig[];
    isTableSettingChangeable?: boolean;
    isMultiSortSettingsEnabled?: boolean;
}

interface ITableSettingsProps {
    isMultiSortEnabled?: boolean;
    isRowCountEnabled?: boolean;
    isSearchEnabled?: boolean;
}

export type { ITableColumnConfig, ITableSettingsProps, ITableConfigProps };