// src/components/Table/models/column.models.ts
import { ITableColumnConfig } from '@/components/Table/models/config.models';
import { EPositionInRow } from '@/components/Table/models/row.models';

type TColumnId = string;

interface IColumnCustomStylingDto {
    align?: EColumnAlign;
    size?: EColumnSize;
}

interface ITableInjectableColumnConfig extends ITableColumnConfig {
    positionInRow: number | EPositionInRow;
}

enum EColumnAlign {
    Left = 'left',
    Right = 'right',
    Center = 'center',
}

enum EColumnSize {
    Xs = 'extra-small',
    Sm = 'small',
    Mid = 'medium',
    Lg = 'large',
    Xl = 'extra-large',
}

enum EColumnDataType {
    Nr = 'nr',
    Number = 'number',
    Date = 'date',
    DateTime = 'datetime',
    String = 'string',
    Text = 'text',
    Unknown = 'unknown',
}

enum EInjectableColumnId {
    RowCount = 'rowCount',
}

const rowCountColumnConfig: ITableInjectableColumnConfig = {
    columnId: EInjectableColumnId.RowCount,
    label: 'Nr',
    dataAccessor: 'rowCount',
    columnDataType: EColumnDataType.Nr,
    isSortable: false,
    isAutoGenerated: true,
    positionInRow: 0,
};

export type { TColumnId, IColumnCustomStylingDto, ITableInjectableColumnConfig };
export { EColumnAlign, EColumnSize, EColumnDataType, EInjectableColumnId, rowCountColumnConfig };